cleanup: Fix logic for implicit subscriptions
Implicit subscriptions should be removed once the respective
implicator no longer holds.

Exemplarily, for a club newsletter, this ensures that people
which became subscribed due to their membership are unsubscribed
once it is terminated.

In doing so, this change removes the implicits_only subscription
policy, which is redundant to subscribable with this change.